+++ lib/gg.dart
@@ -33,6 +33,5 @@ export 'src/commands/did/did_commit.dart';
 // Tools
 export 'src/tools/is_github.dart';
 export 'src/tools/carriage_return.dart';
-export 'src/tools/log_state.dart';
 export 'src/tools/command_cluster.dart';
 export 'src/tools/checks.dart';