+++ lib/src/commands/check/analyze.dart
@@ -7,10 +7,10 @@
 import 'dart:io';
 
 import 'package:gg_args/gg_args.dart';
-import 'package:gg/gg.dart';
 import 'package:gg_log/gg_log.dart';
 import 'package:gg_process/gg_process.dart';
 import 'package:gg_console_colors/gg_console_colors.dart';
+import 'package:gg_status_printer/gg_status_printer.dart';
 import 'package:gg_test/gg_test.dart';
 import 'package:mocktail/mocktail.dart' as mocktail;
 
@@ -33,7 +33,12 @@ class Analyze extends DirCommand<void> {
   }) async {
     await check(directory: directory);
 
-    _logState(LogState.running);
+    final statusPrinter = GgStatusPrinter<ProcessResult>(
+      ggLog: ggLog,
+      message: 'Running "dart analyze"',
+    );
+
+    statusPrinter.logStatus(GgStatusPrinterStatus.running);
 
     final result = await processWrapper.run(
       'dart',
@@ -41,7 +46,11 @@ class Analyze extends DirCommand<void> {
       workingDirectory: directory.path,
     );
 
-    _logState(result.exitCode == 0 ? LogState.success : LogState.error);
+    statusPrinter.logStatus(
+      result.exitCode == 0
+          ? GgStatusPrinterStatus.success
+          : GgStatusPrinterStatus.error,
+    );
 
     if (result.exitCode != 0) {
       final files = [
@@ -61,13 +70,6 @@ class Analyze extends DirCommand<void> {
       );
     }
   }
-
-  // .........................................................................
-  void _logState(LogState state) => logState(
-        ggLog: ggLog,
-        state: state,
-        message: 'Running "dart analyze"',
-      );
 
   /// The process wrapper used to execute shell processes
   final GgProcessWrapper processWrapper;