+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Removed
 
 - Removed unused sample project
+- logStatus is replaced by GgStatusPrinter
 
 ## [1.0.14] - 2024-04-05